pyproject.toml
[project]
name = "fluentra"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "langchain-openai>=1.1.7",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.40.0",
    "cryptography>=40.0.0",
    "asyncpg>=0.31.0",
]
